% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.R
\name{topic_source}
\alias{topic_source}
\title{Find the source package for a topic}
\usage{
topic_source(topic, package)
}
\arguments{
\item{topic}{A single string naming a topic.}

\item{package}{A package name.}
}
\value{
A single package name.
}
\description{
Determines which package supplies the object associated with a documented
topic. This resolves re-exported functions and imported objects to the
package where they originate.
}
\examples{
topic_source("rnorm", "stats")
}
