% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{pkg_cache_reset}
\alias{pkg_cache_reset}
\title{Reset cached package indexes}
\usage{
pkg_cache_reset(package)
}
\arguments{
\item{package}{A package name, or a path to the source directory of a
package.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Clears the cached topic index and parsed Rd objects for \code{package}.
Only needed for source packages, and will generally be automatically
be called by roxygen2.
}
