% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtools-package.R
\docType{package}
\name{rdtools-package}
\alias{rdtools}
\alias{rdtools-package}
\title{rdtools: Efficient Manipulation of 'Rd' Files and Help Topics}
\description{
Fast, cached lookup of help topics and aliases across installed, source, and in-development packages, plus efficient retrieval of parsed 'Rd' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/rdtools}
  \item Report bugs at \url{https://github.com/r-lib/rdtools/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
